feat(#432): name the operator's actual Fly apps in the manual update command
The middleware reads FLY_APP_NAME/FLY_MACHINE_ID and reports them via
/status; the admin page's new server shell contributes the web-ui app's
own name. Detection is positive-only — anything not demonstrably Fly
keeps the generic placeholders.
